Work around unsafe coroutine function detection.
Here's a quote from Guido to back up what has been done:

Usually you shouldn't be checking for couroutine functions directly.
There are too many cases where something returns in fact a coroutine but
you cannot tell by inspecting the thing you call -- you can only tell
once you've called it (using iscoroutine()). Partials are just one
approach -- lambdas are a different one (personally I usually prefer
lambdas over partials). An advantage of using a lambda (or a plain old
function) is that you can decorate it with @coroutine.

(Note that there are a few places in asyncio that check for coroutine
functions -- but always in the negative, in places where callbacks are
taken that should *not* be coroutines. And this test doesn't need to be
perfect -- it's just helpful to call out common anti-patterns early.)

-- Guido van Rossum