Implemented simpler standby routine.
Left concurrency alone and did one thing after the other.

It is slightly less safe, but it might actually work.  The initial
problem was that the temp. ramps went crazy (maybe because the system
was cycling between "STANDBY" and "AMBIENT").  But now that the temp
ramps are un-broken again, even this cycling shouldn't be a problem.